chore: refactoring VAnchor ts interface
