+++ packages/anchors/src/Common.ts
@@ -1,0 +1,214 @@
+import { BigNumber, BigNumberish, ContractTransaction, ethers } from 'ethers';
+import {
+  VAnchor as VAnchorContract,
+  VAnchor__factory,
+  ChainalysisVAnchor as ChainalysisVAnchorContract,
+  DeterministicDeployFactory as DeterministicDeployFactoryContract,
+  VAnchorEncodeInputs__factory,
+  TokenWrapper__factory,
+} from '@webb-tools/contracts';
+import {
+  toHex,
+  Keypair,
+  toFixedHex,
+  Utxo,
+  MerkleTree,
+  median,
+  mean,
+  max,
+  min,
+  randomBN,
+  CircomProvingManager,
+  ProvingManagerSetupInput,
+  MerkleProof,
+  UtxoGenInput,
+  CircomUtxo,
+  FIELD_SIZE,
+  LeafIdentifier,
+} from '@webb-tools/sdk-core';
+import { hexToU8a, u8aToHex, getChainIdType, ZkComponents } from '@webb-tools/utils';
+
+const encoder = (types, values) => {
+  const abiCoder = ethers.utils.defaultAbiCoder;
+  const encodedParams = abiCoder.encode(types, values);
+  return encodedParams.slice(2);
+};
+
+const create2Address = (factoryAddress, saltHex, initCode) => {
+  const create2Addr = ethers.utils.getCreate2Address(factoryAddress, saltHex, ethers.utils.keccak256(initCode));
+  return create2Addr;
+
+}
+type WebbContracts = VAnchorContract | ChainalysisVAnchorContract;
+
+export class WebbBridge {
+  signer: ethers.Signer;
+  contract: WebbContracts;
+
+  constructor(contract: WebbContracts, signer: ethers.Signer) {
+    this.signer = signer;
+    this.contract = contract;
+  }
+
+  public static async generateUTXO(input: UtxoGenInput): Promise<Utxo> {
+    return CircomUtxo.generateUtxo(input);
+  }
+
+  public static createRootsBytes(rootArray: string[]) {
+    let rootsBytes = '0x';
+    for (let i = 0; i < rootArray.length; i++) {
+      rootsBytes += toFixedHex(rootArray[i]).substr(2);
+    }
+    return rootsBytes; // root byte string (32 * array.length bytes)
+  }
+
+  // Convert a hex string to a byte array
+  public static hexStringToByte(str: string) {
+    if (!str) {
+      return new Uint8Array();
+    }
+
+    var a = [];
+    for (var i = 0, len = str.length; i < len; i += 2) {
+      a.push(parseInt(str.substr(i, 2), 16));
+    }
+
+    return new Uint8Array(a);
+  }
+
+  public async setHandler(handlerAddress: string) {
+    const tx = await this.contract.setHandler(
+      handlerAddress,
+      BigNumber.from(await this.contract.getProposalNonce()).add(1)
+    );
+    await tx.wait();
+  }
+
+  public async setSigner(newSigner: ethers.Signer) {
+    const currentChainId = await this.signer.getChainId();
+    const newChainId = await newSigner.getChainId();
+
+    if (currentChainId === newChainId) {
+      this.signer = newSigner;
+      this.contract = this.contract.connect(newSigner);
+      return true;
+    }
+    return false;
+  }
+  public async createResourceId(): Promise<string> {
+    return toHex(
+      this.contract.address + toHex(getChainIdType(await this.signer.getChainId()), 6).substr(2),
+      32
+    );
+  }
+  public async getMinWithdrawalLimitProposalData(
+    _minimalWithdrawalAmount: string
+  ): Promise<string> {
+    const resourceID = await this.createResourceId();
+    const functionSig = ethers.utils
+      .keccak256(ethers.utils.toUtf8Bytes('configureMinimalWithdrawalLimit(uint256,uint32)'))
+      .slice(0, 10)
+      .padEnd(10, '0');
+    const nonce = Number(await this.contract.getProposalNonce()) + 1;
+    return (
+      '0x' +
+      toHex(resourceID, 32).substr(2) +
+      functionSig.slice(2) +
+      toHex(nonce, 4).substr(2) +
+      toFixedHex(_minimalWithdrawalAmount).substr(2)
+    );
+  }
+
+  public async getMaxDepositLimitProposalData(_maximumDepositAmount: string): Promise<string> {
+    const resourceID = await this.createResourceId();
+    const functionSig = ethers.utils
+      .keccak256(ethers.utils.toUtf8Bytes('configureMaximumDepositLimit(uint256,uint32)'))
+      .slice(0, 10)
+      .padEnd(10, '0');
+    const nonce = Number(await this.contract.getProposalNonce()) + 1;
+    return (
+      '0x' +
+      toHex(resourceID, 32).substr(2) +
+      functionSig.slice(2) +
+      toHex(nonce, 4).substr(2) +
+      toFixedHex(_maximumDepositAmount).substr(2)
+    );
+  }
+
+  // Proposal data is used to update linkedAnchors via bridge proposals
+  // on other chains with this anchor's state
+  public async genProposalData(resourceID: string, merkleRoot: string, leafIndex: number): Promise<string> {
+    // If no leaf index passed in, set it to the most recent one.
+    const chainID = getChainIdType(await this.signer.getChainId());
+    const functionSig = ethers.utils
+      .keccak256(ethers.utils.toUtf8Bytes('updateEdge(bytes32,uint32,bytes32)'))
+      .slice(0, 10)
+      .padEnd(10, '0');
+
+    const srcContract = this.contract.address;
+    const srcResourceId =
+      '0x' +
+      toHex(0, 6).substring(2) +
+      toHex(srcContract, 20).substr(2) +
+      toHex(chainID, 6).substr(2);
+    return (
+      '0x' +
+      toHex(resourceID, 32).substr(2) +
+      functionSig.slice(2) +
+      toHex(leafIndex, 4).substr(2) +
+      toHex(merkleRoot, 32).substr(2) +
+      toHex(srcResourceId, 32).substr(2)
+    );
+  }
+
+  public getExtAmount(inputs: Utxo[], outputs: Utxo[], fee: BigNumberish) {
+    return BigNumber.from(fee)
+      .add(outputs.reduce((sum, x) => sum.add(x.amount), BigNumber.from(0)))
+      .sub(inputs.reduce((sum, x) => sum.add(x.amount), BigNumber.from(0)));
+  }
+  public async padUtxos(utxos: Utxo[], maxLen: number): Promise<Utxo[]> {
+    const evmId = await this.signer.getChainId();
+    const chainId = getChainIdType(evmId);
+    const randomKeypair = new Keypair();
+
+    while (utxos.length !== 2 && utxos.length < maxLen) {
+      utxos.push(
+        await CircomUtxo.generateUtxo({
+          curve: 'Bn254',
+          backend: 'Circom',
+          chainId: chainId.toString(),
+          originChainId: chainId.toString(),
+          amount: '0',
+          blinding: hexToU8a(randomBN(31).toHexString()),
+          keypair: randomKeypair,
+        })
+      );
+    }
+    return utxos;
+  }
+
+  // Proposal data is used to update linkedAnchors via bridge proposals
+  // on other chains with this anchor's state
+
+  public async getHandler(): Promise<string> {
+    return this.contract.handler();
+  }
+
+  public async getHandlerProposalData(newHandler: string): Promise<string> {
+    const resourceID = await this.createResourceId();
+    const functionSig = ethers.utils
+      .keccak256(ethers.utils.toUtf8Bytes('setHandler(address,uint32)'))
+      .slice(0, 10)
+      .padEnd(10, '0');
+    const nonce = Number(await this.contract.getProposalNonce()) + 1;
+
+    return (
+      '0x' +
+      toHex(resourceID, 32).substr(2) +
+      functionSig.slice(2) +
+      toHex(nonce, 4).substr(2) +
+      toHex(newHandler, 20).substr(2)
+    );
+  }
+
+}
