+++ packages/anchors/src/VAnchor.ts
@@ -32,6 +32,7 @@ import {
   IVariableAnchorPublicInputs,
 } from '@webb-tools/interfaces';
 import { hexToU8a, u8aToHex, getChainIdType, ZkComponents } from '@webb-tools/utils';
+import { WebbBridge } from './Common';
 
 const encoder = (types, values) => {
   const abiCoder = ethers.utils.defaultAbiCoder;
@@ -60,7 +61,7 @@ export var proofTimeBenchmark = [];
 // It represents a deployed contract throughout its life (e.g. maintains merkle tree state)
 // Functionality relevant to anchors in general (proving, verifying) is implemented in static methods
 // Functionality relevant to a particular anchor deployment (deposit, withdraw) is implemented in instance methods
-export class VAnchor implements IAnchor {
+export class VAnchor extends WebbBridge implements IAnchor {
   signer: ethers.Signer;
   contract: VAnchorContract | ChainalysisVAnchorContract;
   tree: MerkleTree;
@@ -84,6 +85,7 @@ export class VAnchor implements IAnchor {
     smallCircuitZkComponents: ZkComponents,
     largeCircuitZkComponents: ZkComponents
   ) {
+    super(contract, signer)
     this.signer = signer;
     this.contract = contract;
     this.tree = new MerkleTree(treeHeight);
@@ -117,8 +119,6 @@ export class VAnchor implements IAnchor {
     const factory1Create2Addr = create2Address(deployer.address, saltHex, encodeLibrary1InitCode)
     const encodeLibrary1Tx = await deployer.deploy(encodeLibrary1InitCode, saltHex);
     const encodeLibrary1Receipt = await encodeLibrary1Tx.wait()
-    console.log("factory receipt", encodeLibrary1Receipt.events)
-    // const contract1 = await encodeLibrary1Factory.attach(encodeLibrary1Receipt.events[encodeLibrary1Receipt.events.length - 1].args[0]);
     let libraryAddress = encodeLibrary1Receipt.events[encodeLibrary1Receipt.events.length - 1].args[0]
     const factory = new VAnchor__factory(
       { ['contracts/libs/VAnchorEncodeInputs.sol:VAnchorEncodeInputs']: libraryAddress },
@@ -128,12 +128,7 @@ export class VAnchor implements IAnchor {
     const vanchorInitCode = vanchorBytecode + encoder(["address", "uint32", "address", "address", "address", "uint8"], [verifier, levels, hasher, handler, token, maxEdges])
     const vanchorTx = await deployer.deploy(vanchorInitCode, saltHex);
     const vanchorReceipt = await vanchorTx.wait()
-    console.log("vanchor receipt", vanchorReceipt.events[0])
-    console.log("Vanchor deployed to ", vanchorReceipt.events[0].args[0])
     const vanchor = await factory.attach(vanchorReceipt.events[0].args[0]);
-    // const vAnchor = await factory.deploy(verifier, levels, hasher, handler, token, maxEdges, {});
-    console.log("vanchor obj", vanchor)
-    // await vAnchor.deployed();
     const createdVAnchor = new VAnchor(
       vanchor,
       signer,
@@ -203,31 +198,6 @@ export class VAnchor implements IAnchor {
     return createdAnchor;
   }
 
-  public static async generateUTXO(input: UtxoGenInput): Promise<Utxo> {
-    return CircomUtxo.generateUtxo(input);
-  }
-
-  public static createRootsBytes(rootArray: string[]) {
-    let rootsBytes = '0x';
-    for (let i = 0; i < rootArray.length; i++) {
-      rootsBytes += toFixedHex(rootArray[i]).substr(2);
-    }
-    return rootsBytes; // root byte string (32 * array.length bytes)
-  }
-
-  // Convert a hex string to a byte array
-  public static hexStringToByte(str: string) {
-    if (!str) {
-      return new Uint8Array();
-    }
-
-    var a = [];
-    for (var i = 0, len = str.length; i < len; i += 2) {
-      a.push(parseInt(str.substr(i, 2), 16));
-    }
-
-    return new Uint8Array(a);
-  }
 
   public static convertToPublicInputsStruct(args: any[]): IVariableAnchorPublicInputs {
     return {
@@ -237,19 +207,6 @@ export class VAnchor implements IAnchor {
       outputCommitments: args[3],
       publicAmount: args[4],
       extDataHash: args[5],
-    };
-  }
-
-  public static convertToExtDataStruct(args: any[]): IVariableAnchorExtData {
-    return {
-      recipient: args[0],
-      extAmount: args[1],
-      relayer: args[2],
-      fee: args[3],
-      refund: args[4],
-      token: args[5],
-      encryptedOutput1: args[6],
-      encryptedOutput2: args[7],
     };
   }
 
@@ -264,39 +221,12 @@ export class VAnchor implements IAnchor {
     // this.latestSyncedBlock = currentBlockNumber;
   }
 
-  public async createResourceId(): Promise<string> {
-    return toHex(
-      this.contract.address + toHex(getChainIdType(await this.signer.getChainId()), 6).substr(2),
-      32
-    );
-  }
-
   public async setVerifier(verifierAddress: string) {
     const tx = await this.contract.setVerifier(
       verifierAddress,
       BigNumber.from(await this.contract.getProposalNonce()).add(1)
     );
     await tx.wait();
-  }
-
-  public async setHandler(handlerAddress: string) {
-    const tx = await this.contract.setHandler(
-      handlerAddress,
-      BigNumber.from(await this.contract.getProposalNonce()).add(1)
-    );
-    await tx.wait();
-  }
-
-  public async setSigner(newSigner: ethers.Signer) {
-    const currentChainId = await this.signer.getChainId();
-    const newChainId = await newSigner.getChainId();
-
-    if (currentChainId === newChainId) {
-      this.signer = newSigner;
-      this.contract = this.contract.connect(newSigner);
-      return true;
-    }
-    return false;
   }
 
   // Proposal data is used to update linkedAnchors via bridge proposals
@@ -307,82 +237,8 @@ export class VAnchor implements IAnchor {
       leafIndex = this.tree.number_of_elements() - 1;
     }
 
-    const chainID = getChainIdType(await this.signer.getChainId());
     const merkleRoot = this.depositHistory[leafIndex];
-    const functionSig = ethers.utils
-      .keccak256(ethers.utils.toUtf8Bytes('updateEdge(bytes32,uint32,bytes32)'))
-      .slice(0, 10)
-      .padEnd(10, '0');
-
-    const srcContract = this.contract.address;
-    const srcResourceId =
-      '0x' +
-      toHex(0, 6).substring(2) +
-      toHex(srcContract, 20).substr(2) +
-      toHex(chainID, 6).substr(2);
-    return (
-      '0x' +
-      toHex(resourceID, 32).substr(2) +
-      functionSig.slice(2) +
-      toHex(leafIndex, 4).substr(2) +
-      toHex(merkleRoot, 32).substr(2) +
-      toHex(srcResourceId, 32).substr(2)
-    );
-  }
-
-  public async getHandler(): Promise<string> {
-    return this.contract.handler();
-  }
-
-  public async getHandlerProposalData(newHandler: string): Promise<string> {
-    const resourceID = await this.createResourceId();
-    const functionSig = ethers.utils
-      .keccak256(ethers.utils.toUtf8Bytes('setHandler(address,uint32)'))
-      .slice(0, 10)
-      .padEnd(10, '0');
-    const nonce = Number(await this.contract.getProposalNonce()) + 1;
-
-    return (
-      '0x' +
-      toHex(resourceID, 32).substr(2) +
-      functionSig.slice(2) +
-      toHex(nonce, 4).substr(2) +
-      toHex(newHandler, 20).substr(2)
-    );
-  }
-
-  public async getMinWithdrawalLimitProposalData(
-    _minimalWithdrawalAmount: string
-  ): Promise<string> {
-    const resourceID = await this.createResourceId();
-    const functionSig = ethers.utils
-      .keccak256(ethers.utils.toUtf8Bytes('configureMinimalWithdrawalLimit(uint256,uint32)'))
-      .slice(0, 10)
-      .padEnd(10, '0');
-    const nonce = Number(await this.contract.getProposalNonce()) + 1;
-    return (
-      '0x' +
-      toHex(resourceID, 32).substr(2) +
-      functionSig.slice(2) +
-      toHex(nonce, 4).substr(2) +
-      toFixedHex(_minimalWithdrawalAmount).substr(2)
-    );
-  }
-
-  public async getMaxDepositLimitProposalData(_maximumDepositAmount: string): Promise<string> {
-    const resourceID = await this.createResourceId();
-    const functionSig = ethers.utils
-      .keccak256(ethers.utils.toUtf8Bytes('configureMaximumDepositLimit(uint256,uint32)'))
-      .slice(0, 10)
-      .padEnd(10, '0');
-    const nonce = Number(await this.contract.getProposalNonce()) + 1;
-    return (
-      '0x' +
-      toHex(resourceID, 32).substr(2) +
-      functionSig.slice(2) +
-      toHex(nonce, 4).substr(2) +
-      toFixedHex(_maximumDepositAmount).substr(2)
-    );
+    return this.genProposalData(resourceID, merkleRoot, leafIndex);
   }
 
   public async populateRootsForProof(): Promise<string[]> {
@@ -509,6 +365,18 @@ export class VAnchor implements IAnchor {
       minTime,
     };
   }
+  public static convertToExtDataStruct(args: any[]): IVariableAnchorExtData {
+    return {
+      recipient: args[0],
+      extAmount: args[1],
+      relayer: args[2],
+      fee: args[3],
+      refund: args[4],
+      token: args[5],
+      encryptedOutput1: args[6],
+      encryptedOutput2: args[7],
+    };
+  }
 
   /**
    *
@@ -517,8 +385,7 @@ export class VAnchor implements IAnchor {
    */
   public async setupTransaction(
     inputs: Utxo[],
-    outputs: [Utxo, Utxo],
-    extAmount: BigNumberish,
+    outputs: Utxo[],
     fee: BigNumberish,
     refund: BigNumberish,
     token: string,
@@ -526,9 +393,9 @@ export class VAnchor implements IAnchor {
     relayer: string,
     leavesMap: Record<string, Uint8Array[]>
   ) {
-    // first, check if the merkle root is known on chain - if not, then update
-    const chainId = getChainIdType(await this.signer.getChainId());
-    const roots = await this.populateRootsForProof();
+    // Default UTXO chain ID will match with the configured signer's chain ID
+    inputs = await this.padUtxos(inputs, 16)
+    outputs = await this.padUtxos(outputs, 2)
 
     // calculate the sum of input notes (for calculating the public amount)
     let sumInputUtxosAmount: BigNumberish = 0;
@@ -544,43 +411,9 @@ export class VAnchor implements IAnchor {
       });
     }
 
-    const encryptedCommitments: [Uint8Array, Uint8Array] = [
-      hexToU8a(outputs[0].encrypt()),
-      hexToU8a(outputs[1].encrypt()),
-    ];
+    const roots = await this.populateRootsForProof();
 
-    const proofInput: ProvingManagerSetupInput<'vanchor'> = {
-      inputUtxos: inputs,
-      leavesMap,
-      leafIds,
-      roots: roots.map((root) => hexToU8a(root)),
-      chainId: chainId.toString(),
-      output: outputs,
-      encryptedCommitments,
-      publicAmount: BigNumber.from(extAmount).sub(fee).add(FIELD_SIZE).mod(FIELD_SIZE).toString(),
-      provingKey:
-        inputs.length > 2 ? this.largeCircuitZkComponents.zkey : this.smallCircuitZkComponents.zkey,
-      relayer: hexToU8a(relayer),
-      recipient: hexToU8a(recipient),
-      extAmount: toFixedHex(BigNumber.from(extAmount)),
-      fee: BigNumber.from(fee).toString(),
-      refund: BigNumber.from(refund).toString(),
-      token: hexToU8a(token),
-    };
-
-    inputs.length > 2
-      ? (this.provingManager = new CircomProvingManager(
-        this.largeCircuitZkComponents.wasm,
-        this.tree.levels,
-        null
-      ))
-      : (this.provingManager = new CircomProvingManager(
-        this.smallCircuitZkComponents.wasm,
-        this.tree.levels,
-        null
-      ));
-
-    const proof = await this.provingManager.prove('vanchor', proofInput);
+    const { extAmount, proof, proofInput } = await this.generateProof(roots, inputs, outputs, fee, refund, token, recipient, relayer, leafIds, leavesMap)
 
     const publicInputs: IVariableAnchorPublicInputs = this.generatePublicInputs(
       proof.proof,
@@ -603,6 +436,7 @@ export class VAnchor implements IAnchor {
     };
 
     return {
+      extAmount,
       extData,
       publicInputs,
     };
@@ -618,59 +452,13 @@ export class VAnchor implements IAnchor {
     relayer: string
   ): Promise<ethers.ContractReceipt> {
     // Validate input utxos have a valid originChainId
-    inputs.map((utxo) => {
-      if (utxo.originChainId === undefined) {
-        throw new Error('Input Utxo does not have a configured originChainId');
-      }
-    });
-
-    // Default UTXO chain ID will match with the configured signer's chain ID
-    const evmId = await this.signer.getChainId();
-    const chainId = getChainIdType(evmId);
-    const randomKeypair = new Keypair();
-
-    while (inputs.length !== 2 && inputs.length < 16) {
-      inputs.push(
-        await CircomUtxo.generateUtxo({
-          curve: 'Bn254',
-          backend: 'Circom',
-          chainId: chainId.toString(),
-          originChainId: chainId.toString(),
-          amount: '0',
-          blinding: hexToU8a(randomBN(31).toHexString()),
-          keypair: randomKeypair,
-        })
-      );
-    }
-
-    if (outputs.length < 2) {
-      while (outputs.length < 2) {
-        outputs.push(
-          await CircomUtxo.generateUtxo({
-            curve: 'Bn254',
-            backend: 'Circom',
-            chainId: chainId.toString(),
-            originChainId: chainId.toString(),
-            amount: '0',
-            keypair: randomKeypair,
-          })
-        );
-      }
-    }
-
-    let extAmount = BigNumber.from(fee)
-      .add(outputs.reduce((sum, x) => sum.add(x.amount), BigNumber.from(0)))
-      .sub(inputs.reduce((sum, x) => sum.add(x.amount), BigNumber.from(0)));
-
-    const token = this.token;
-
+    this.validateInputs(inputs);
     const { extData, publicInputs } = await this.setupTransaction(
       inputs,
-      [outputs[0], outputs[1]],
-      extAmount,
+      outputs,
       fee,
       refund,
-      token,
+      this.token,
       recipient,
       relayer,
       leavesMap
@@ -688,13 +476,24 @@ export class VAnchor implements IAnchor {
     gasBenchmark.push(receipt.gasUsed.toString());
 
     // Add the leaves to the tree
+    this.updateTreeState(outputs);
+
+    return receipt;
+  }
+
+  public updateTreeState(outputs: Utxo[]): void {
     outputs.forEach((x) => {
       this.tree.insert(u8aToHex(x.commitment));
       let numOfElements = this.tree.number_of_elements();
       this.depositHistory[numOfElements - 1] = toFixedHex(this.tree.root().toString());
     });
-
-    return receipt;
+  }
+  public validateInputs(inputs: Utxo[]): void {
+    inputs.map((utxo) => {
+      if (utxo.originChainId === undefined) {
+        throw new Error('Input Utxo does not have a configured originChainId');
+      }
+    });
   }
 
   public async transactWrap(
@@ -707,48 +506,10 @@ export class VAnchor implements IAnchor {
     relayer: string,
     leavesMap: Record<string, Uint8Array[]>
   ): Promise<ethers.ContractReceipt> {
-    // Default UTXO chain ID will match with the configured signer's chain ID
-    const evmId = await this.signer.getChainId();
-    const chainId = getChainIdType(evmId);
-    const randomKeypair = new Keypair();
-
-    while (inputs.length !== 2 && inputs.length < 16) {
-      inputs.push(
-        await CircomUtxo.generateUtxo({
-          curve: 'Bn254',
-          backend: 'Circom',
-          chainId: chainId.toString(),
-          originChainId: chainId.toString(),
-          amount: '0',
-          blinding: hexToU8a(randomBN(31).toHexString()),
-          keypair: randomKeypair,
-        })
-      );
-    }
-
-    if (outputs.length < 2) {
-      while (outputs.length < 2) {
-        outputs.push(
-          await CircomUtxo.generateUtxo({
-            curve: 'Bn254',
-            backend: 'Circom',
-            chainId: chainId.toString(),
-            originChainId: chainId.toString(),
-            amount: '0',
-            keypair: randomKeypair,
-          })
-        );
-      }
-    }
-
-    let extAmount = BigNumber.from(fee)
-      .add(outputs.reduce((sum, x) => sum.add(x.amount), BigNumber.from(0)))
-      .sub(inputs.reduce((sum, x) => sum.add(x.amount), BigNumber.from(0)));
-
-    const { extData, publicInputs } = await this.setupTransaction(
+    // TODO: VERIFY IF WE SHOULD VALIDATE INPUTS HERE
+    const { extAmount, extData, publicInputs } = await this.setupTransaction(
       inputs,
-      [outputs[0], outputs[1]],
-      extAmount,
+      outputs,
       fee,
       refund,
       tokenAddress,
@@ -788,14 +549,64 @@ export class VAnchor implements IAnchor {
     const receipt = await tx.wait();
 
     // Add the leaves to the tree
-    outputs.forEach((x) => {
-      // Maintain tree state after insertions
-      this.tree.insert(u8aToHex(x.commitment));
-      let numOfElements = this.tree.number_of_elements();
-      this.depositHistory[numOfElements - 1] = toFixedHex(this.tree.root().toString());
-    });
+    this.updateTreeState(outputs);
 
     return receipt;
+  }
+
+  public async generateProof(
+    roots: string[],
+    inputs: Utxo[],
+    outputs: Utxo[],
+    fee: BigNumberish,
+    refund: BigNumberish,
+    tokenAddress: string,
+    recipient: string,
+    relayer: string,
+    leafIds: LeafIdentifier[],
+    leavesMap: Record<string, Uint8Array[]>
+  ) {
+    let extAmount = this.getExtAmount(inputs, outputs, fee)
+
+    const encryptedCommitments: [Uint8Array, Uint8Array] = [
+      hexToU8a(outputs[0].encrypt()),
+      hexToU8a(outputs[1].encrypt()),
+    ];
+    const chainId = getChainIdType(await this.signer.getChainId());
+
+    const proofInput: ProvingManagerSetupInput<'vanchor'> = {
+      inputUtxos: inputs,
+      leavesMap,
+      leafIds,
+      roots: roots.map((root) => hexToU8a(root)),
+      chainId: chainId.toString(),
+      output: [outputs[0], outputs[1]],
+      encryptedCommitments,
+      publicAmount: BigNumber.from(extAmount).sub(fee).add(FIELD_SIZE).mod(FIELD_SIZE).toString(),
+      provingKey:
+        inputs.length > 2 ? this.largeCircuitZkComponents.zkey : this.smallCircuitZkComponents.zkey,
+      relayer: hexToU8a(relayer),
+      recipient: hexToU8a(recipient),
+      extAmount: toFixedHex(BigNumber.from(extAmount)),
+      fee: BigNumber.from(fee).toString(),
+      refund: BigNumber.from(refund).toString(),
+      token: hexToU8a(tokenAddress),
+    };
+
+    inputs.length > 2
+      ? (this.provingManager = new CircomProvingManager(
+        this.largeCircuitZkComponents.wasm,
+        this.tree.levels,
+        null
+      ))
+      : (this.provingManager = new CircomProvingManager(
+        this.smallCircuitZkComponents.wasm,
+        this.tree.levels,
+        null
+      ));
+
+    const proof = await this.provingManager.prove('vanchor', proofInput);
+    return { proof, extAmount, proofInput }
   }
 
   public async registerAndTransact(
@@ -809,52 +620,14 @@ export class VAnchor implements IAnchor {
     relayer: string,
     leavesMap: Record<string, Uint8Array[]>
   ): Promise<ethers.ContractReceipt> {
-    const chainId = getChainIdType(await this.signer.getChainId());
-    const randomKeypair = new Keypair();
-
-    while (inputs.length !== 2 && inputs.length < 16) {
-      inputs.push(
-        await CircomUtxo.generateUtxo({
-          curve: 'Bn254',
-          backend: 'Circom',
-          chainId: chainId.toString(),
-          originChainId: chainId.toString(),
-          blinding: hexToU8a(randomBN(31).toHexString()),
-          amount: '0',
-          keypair: randomKeypair,
-        })
-      );
-    }
-
-    if (outputs.length < 2) {
-      while (outputs.length < 2) {
-        outputs.push(
-          await CircomUtxo.generateUtxo({
-            curve: 'Bn254',
-            backend: 'Circom',
-            chainId: chainId.toString(),
-            originChainId: chainId.toString(),
-            blinding: hexToU8a(randomBN(31).toHexString()),
-            amount: '0',
-            keypair: randomKeypair,
-          })
-        );
-      }
-    }
-
-    let extAmount = BigNumber.from(fee)
-      .add(outputs.reduce((sum, x) => sum.add(BigNumber.from(BigInt(x.amount))), BigNumber.from(0)))
-      .sub(inputs.reduce((sum, x) => sum.add(BigNumber.from(BigInt(x.amount))), BigNumber.from(0)));
-
-    const token = this.token;
+    // TODO: VERIFY IF WE SHOULD VALIDATE INPUTS HERE
 
     const { extData, publicInputs } = await this.setupTransaction(
       inputs,
-      [outputs[0], outputs[1]],
-      extAmount,
+      outputs,
       fee,
       refund,
-      token,
+      this.token,
       recipient,
       relayer,
       leavesMap
@@ -872,11 +645,7 @@ export class VAnchor implements IAnchor {
     const receipt = await tx.wait();
 
     // Add the leaves to the tree
-    outputs.forEach((x) => {
-      this.tree.insert(u8aToHex(x.commitment));
-      let numOfElements = this.tree.number_of_elements();
-      this.depositHistory[numOfElements - 1] = toFixedHex(this.tree.root().toString());
-    });
+    this.updateTreeState(outputs);
 
     return receipt;
   }
